+++ src/redux/store.js
@@ -3,12 +3,14 @@ import logger from 'redux-logger';
 import thunk from 'redux-thunk';
 import authenticationSlice from './authentication/authenticationSlice';
 import registrationSlice from './registration/registrationSlice';
+import carSlice from './car/carSlice';
 
 const store = configureStore({
   middleware: [logger, thunk],
   reducer: {
     authentication: authenticationSlice,
     registration: registrationSlice,
+    car: carSlice,
   },
 });
